requirements.txt
requests
tqdm
pandas
dask
pyarrow
jupyterlab
fonttools
Pillow
dvc
pyclipper
label_studio
cjkwrap